Major additions to setup, also added Roku functionality
RBP Setup has been improved, so now it should work out of the box.
Additionally, we now have a script to setup the environment variables, and
initialize the config/ directory with appropriate default configurations.

The Roku device now supports playing a random episode from a show or set
of shows (given by the config file), as well as playing specific content
given by the config file.